# swagger.yaml
openapi: 3.0.0
info:
  title: Hono BE API
  description: >-
    API Documentation for Hono BE Backend Service.
    Language is detected from the `Accept-Language` header (`en` default, `id` supported)
    and echoed back via the `Content-Language` response header on every response.
  version: 2.0.0
servers:
  - url: http://localhost:4000/api
    description: Local development server
  - url: /api
    description: Relative API server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. Expires in 15 minutes."
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: >-
        Preferred response language. Supported: `en` (default), `id`.
        Missing or unsupported values fall back to `en`.
      schema:
        type: string
        enum: [en, id]
        default: en
      example: id
...
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        statusCode: { type: integer, example: 200 }
        message: { type: string, example: "Operation successful" }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        statusCode: { type: integer, example: 400 }
        message: { type: string, example: "Something went wrong" }
        errors: { type: object, nullable: true, example: null }
...
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, example: "Budi Santoso" }
        email: { type: string, format: email, example: "budi@example.com" }
        password: { type: string, minLength: 6, example: "secret123" }
...
    AuthUser:
      type: object
      properties:
        id: { type: integer, example: 1 }
        name: { type: string, example: "Budi Santoso" }
        photo: { type: string, nullable: true, example: "https://cdn.example.com/photo.jpg" }
        email: { type: string, format: email, example: "budi@example.com" }
        isActive: { type: boolean, example: true }
        hasPassword:
          type: boolean
          description: Whether the account has a password set (false for Google-only accounts)
          example: true
...
paths:
...
  /auth/register:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthUser" }
        "400":
          description: Email already in use
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }

  # ── Login ─────────────────────────────────────────────────────────────────
  /auth/login:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Logged in successfully
          headers:
            Content-Language: { $ref: "#/components/headers/ContentLanguage" }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResponse" }
        "401":
          description: User not registered, account inactive, or invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }

  # ── Google Login ──────────────────────────────────────────────────────────
  /auth/google:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    post:
      tags: [Auth]
      summary: Login with Google OAuth
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GoogleLoginRequest" }
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResponse" }
        "400":
          description: User not registered or account inactive
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }

  # ── Forgot Password ──────────────────────────────────────────────────────
  /auth/forgot-password:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    post:
      tags: [Auth]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForgotPasswordRequest" }
      responses:
        "200":
          description: Password reset instructions sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BaseResponse" }
        "400":
          description: Email not found or account inactive
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }

  # ── Validate Reset Token ─────────────────────────────────────────────────
  /auth/validate-reset-token:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [Auth]
      summary: Validate a password reset token
      parameters:
        - name: email
          in: query
          required: true
          schema: { type: string, format: email }
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Token is valid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BaseResponse" }
        "400":
          description: Missing email/token, or invalid/expired reset token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ── Reset Password ──────────────────────────────────────────────────────
  /auth/reset-password:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    post:
      tags: [Auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordRequest" }
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BaseResponse" }
        "400":
          description: Invalid or expired reset token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }

  # ── Refresh Token ────────────────────────────────────────────────────────
  /auth/refresh:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshTokenRequest" }
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResponse" }
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }

  # ── Me ───────────────────────────────────────────────────────────────────
  /auth/me:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [Auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthUser" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/profile:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    put:
      tags: [Auth]
      summary: Update current user's profile details
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProfileRequest" }
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthUser" }
        "400":
          description: Email already in use
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }
  /auth/password:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    put:
      tags: [Auth]
      summary: Update current user's password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePasswordRequest" }
      responses:
        "200":
          description: Password updated successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BaseResponse" }
        "400":
          description: Invalid old password, or old password required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }

  # ── Logout ───────────────────────────────────────────────────────────────
  /auth/logout:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    post:
      tags: [Auth]
      summary: Logout current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BaseResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ── User List ────────────────────────────────────────────────────────────
  /user:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [User]
      summary: Get all users (paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: q
          in: query
          description: Search by name or email
          schema: { type: string }
        - name: isActive
          in: query
          description: Filter by status ("1" = active, "0" = inactive)
          schema: { type: string, enum: ["1", "0"] }
        - name: sortBy
          in: query
          description: "Column to sort by. Falls back to id if omitted or not sortable."
          schema: { type: string, enum: [name, email, isActive, createdAt] }
        - name: order
          in: query
          schema: { type: string, enum: [ASC, DESC], default: DESC }
      responses:
        "200":
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/User" }
                      meta: { $ref: "#/components/schemas/PaginationMeta" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/list:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [User]
      summary: Get all users as a flat list (for select dropdown options)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users list retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: integer, example: 1 }
                            name: { type: string, example: "John Doe" }
                            photo: { type: string, nullable: true, example: "https://minio/avatar.jpg" }
                            email: { type: string, nullable: true, example: "john@example.com" }
                            isActive: { type: boolean, example: true }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    post:
      tags: [User]
      summary: Create a new user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "400":
          description: Email already in use
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }

  # ── User Detail / Update / Delete ────────────────────────────────────────
  /user/{id}:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [User]
      summary: Get user by ID
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: User not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    put:
      tags: [User]
      summary: Update a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRequest" }
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "400":
          description: Email already in use
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: User not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }

    delete:
      tags: [User]
      summary: Delete a user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User deleted successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BaseResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: User not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ── Content Categories ───────────────────────────────────────────────────
  /content/categories:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [Content - Categories]
      summary: Get all content categories (paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: q
          in: query
          description: Search by name or description
          schema: { type: string }
        - name: sortBy
          in: query
          description: Column to sort by
          schema: { type: string, enum: [name, createdAt] }
        - name: order
          in: query
          schema: { type: string, enum: [ASC, DESC], default: DESC }
      responses:
        "200":
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Category" }
                      meta: { $ref: "#/components/schemas/PaginationMeta" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    post:
      tags: [Content - Categories]
      summary: Create a new category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCategoryRequest" }
      responses:
        "201":
          description: Category created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Category" }
        "400":
          description: Category slug already in use or invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }
  /content/categories/all:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [Content - Categories]
      summary: Get all categories as an unpaginated list (for dropdowns)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Category" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /content/categories/{id}:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Content - Categories]
      summary: Get category by ID
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Category retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Category" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Category not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    put:
      tags: [Content - Categories]
      summary: Update category by ID
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCategoryRequest" }
      responses:
        "200":
          description: Category updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Category" }
        "400":
          description: Category slug already in use or invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Category not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }

    delete:
      tags: [Content - Categories]
      summary: Delete category by ID
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Category deleted successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BaseResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Category not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ── Content Articles ─────────────────────────────────────────────────────
...
  /content/articles/slug/{slug}:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
      - name: slug
        in: path
        required: true
        description: Unique article URL slug
        schema: { type: string }
    get:
      tags: [Content - Articles]
      summary: Get article by slug (Public endpoint)
      responses:
        "200":
          description: Article retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Article" }
        "404":
          description: Article not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...